Fix display of the examples/starters on small screens
